+++ modules/graphics/src/include/RenderInterface.hpp
@@ -189,6 +189,12 @@ public:
         = 0;
     //=============================================================================
     virtual void
+    drawImage(const std::vector<double>& xp, const std::vector<double>& yp,
+        const std::vector<double>& xLim, const std::vector<double>& yLim, bool xFlip, bool yFlip,
+        QImage pic)
+        = 0;
+    //=============================================================================
+    virtual void
     quadStrips(std::vector<std::vector<coloredPoint>> faces, bool flatfaces,
         std::vector<std::vector<coloredPoint>> edges, bool flatedges)
         = 0;